# Page#dir: ensure it ends in a slash