Fixes #3974.

/cc @envygeeks, because I'm sure you have a faster/more elegant/more performant solution for this problem.
/cc @jekyll/core for review!