@envygeeks 2016-01-27
```ruby
def _new(url)
  return url if url.end_with?("/") || url == "/"

  url = File.dirname(url)
  url == "/" ? url : "#{url}/"
end
```

```
Calculating -------------------------------------
              _old:/    60.796k i/100ms
              _new:/    72.437k i/100ms
-------------------------------------------------
              _old:/      2.489M (± 4.8%) i/s -     12.463M
              _new:/      4.657M (± 8.5%) i/s -     23.107M

Comparison:
              _new:/:  4657430.1 i/s
              _old:/:  2489307.2 i/s - 1.87x slower

Calculating -------------------------------------
   _old:/hello/world    43.174k i/100ms
   _new:/hello/world    46.844k i/100ms
-------------------------------------------------
   _old:/hello/world      1.172M (± 3.5%) i/s -      5.872M
   _new:/hello/world      1.261M (± 4.0%) i/s -      6.324M

Comparison:
   _new:/hello/world:  1261134.4 i/s
   _old:/hello/world:  1172086.2 i/s - 1.08x slower

Calculating -------------------------------------
  _old:/hello/world/    61.655k i/100ms
  _new:/hello/world/    72.840k i/100ms
-------------------------------------------------
  _old:/hello/world/      2.531M (± 3.4%) i/s -     12.639M
  _new:/hello/world/      4.752M (± 4.0%) i/s -     23.746M

Comparison:
  _new:/hello/world/:  4751972.9 i/s
  _old:/hello/world/:  2531102.8 i/s - 1.88x slower

```

---

You can edge out more performance with `url + "/"` but that's a violation of Github's styleguide which I follow so I did not include it in the results.