@alfredxing 2016-01-27
:+1: for @envygeeks' function. Though I think the `|| url == "/"` can be removed from the first line? (Since `.end_with? "/"` should return `true` in this case)