@alfredxing 2016-01-27
OK! But (just a sanity check) doesn't the first part (`url.end_with?("/")`) get evaluated first, then `url == "/"`?