@envygeeks 2016-01-27
It should but sadly the code ran faster with a more direct check, in there too.  I'll double check the benches again just to make sure.