@parkr 2016-01-27
Ok, I just pushed up another iteration. A bit longer, but doing every little thing helped:

```text
Calculating -------------------------------------
            pre_pr:/   151.086k i/100ms
                pr:/   168.945k i/100ms
         envygeeks:/   168.146k i/100ms
-------------------------------------------------
            pre_pr:/      5.652M (± 4.5%) i/s -     28.253M
                pr:/      8.017M (± 5.4%) i/s -     40.040M
         envygeeks:/      7.445M (±10.1%) i/s -     36.824M

Comparison:
                pr:/:  8017279.2 i/s
         envygeeks:/:  7444605.1 i/s - 1.08x slower
            pre_pr:/:  5651646.7 i/s - 1.42x slower

Calculating -------------------------------------
pre_pr:/some/very/very/long/path/to/a/file/i/like
                        87.137k i/100ms
pr:/some/very/very/long/path/to/a/file/i/like
                        62.901k i/100ms
envygeeks:/some/very/very/long/path/to/a/file/i/like
                        60.416k i/100ms
-------------------------------------------------
pre_pr:/some/very/very/long/path/to/a/file/i/like
                          1.396M (± 6.4%) i/s -      6.971M
pr:/some/very/very/long/path/to/a/file/i/like
                        895.033k (± 5.4%) i/s -      4.466M
envygeeks:/some/very/very/long/path/to/a/file/i/like
                        891.898k (± 4.9%) i/s -      4.471M

Comparison:
pre_pr:/some/very/very/long/path/to/a/file/i/like:  1396111.2 i/s
pr:/some/very/very/long/path/to/a/file/i/like:   895033.2 i/s - 1.56x slower
envygeeks:/some/very/very/long/path/to/a/file/i/like:   891897.8 i/s - 1.57x slower

Calculating -------------------------------------
pre_pr:/some/very/very/long/path/to/a/file/i/like/
                       150.546k i/100ms
pr:/some/very/very/long/path/to/a/file/i/like/
                       165.396k i/100ms
envygeeks:/some/very/very/long/path/to/a/file/i/like/
                       162.486k i/100ms
-------------------------------------------------
pre_pr:/some/very/very/long/path/to/a/file/i/like/
                          5.570M (± 5.8%) i/s -     27.851M
pr:/some/very/very/long/path/to/a/file/i/like/
                          7.821M (± 7.1%) i/s -     39.033M
envygeeks:/some/very/very/long/path/to/a/file/i/like/
                          7.734M (± 7.6%) i/s -     38.509M

Comparison:
pr:/some/very/very/long/path/to/a/file/i/like/:  7821419.0 i/s
envygeeks:/some/very/very/long/path/to/a/file/i/like/:  7733559.1 i/s - 1.01x slower
pre_pr:/some/very/very/long/path/to/a/file/i/like/:  5569945.8 i/s - 1.40x slower
```