@parkr 2016-01-27
@jekyllbot: merge +bug